Add cloudwatch (#66)
* Add cloudwatch logging

Fix rebase

Handle mix of timeout and results (#65)

* Handle mix of timeout and results

* Fix removed line

Fix gemfile.lock

* Fix Gemfile.lock

* Refactor (#67)

* Reshape repo

* Add canary checks

* Add canary checks

* Add some tests

* Add support for testing on non-main branch

* Add support for x86 in tests
