+++ lib/tooling_invoker/job_processor/process_job.rb
@@ -1,0 +1,30 @@
+module ToolingInvoker
+  module JobProcessor
+    class ProcessJob
+      include Mandate
+
+      initialize_with :job
+
+      def call
+        Log.("Invoking request: #{job.type}:#{job.language}:#{job.exercise}", job:)
+
+        JobProcessor::PrepareInput.(job) && run_job!
+      end
+
+      private
+      def run_job!
+        start_time = Time.now.to_f
+
+        begin
+          Log.("Invoking container", job:)
+
+          ExecDocker.(job)
+        rescue StandardError => e
+          job.exceptioned!(e.message, backtrace: e.backtrace)
+        ensure
+          job.duration = Time.now.to_f - start_time
+        end
+      end
+    end
+  end
+end
