+++ lib/tooling_invoker/worker.rb
@@ -2,6 +2,9 @@ module ToolingInvoker
   class Worker
     extend Mandate::InitializerInjector
 
+    class ExitWorkerError < StandardError
+    end
+
     initialize_with :worker_idx
 
     def exit!
@@ -26,15 +29,18 @@ module ToolingInvoker
           counter += 1
         end
 
-        job = check_for_job
+        job = Worker::RetrieveJob.()
+
         if job
           Log.("Starting job", job:)
           start_time = Time.now.to_f
-          handle_job(job)
+          Worker::HandleJob.(job)
           Log.("Total time: #{Time.now.to_f - start_time}", job:)
         else
           sleep(ToolingInvoker.config.job_polling_delay)
         end
+      rescue ExitWorkerError
+        exit!
       rescue StandardError => e
         Log.("Top level error")
         Log.(e.message)
@@ -47,52 +53,6 @@ module ToolingInvoker
     private
     def should_exit?
       @should_exit
-    end
-
-    # This will raise an exception if something other than
-    # a 200 or a 404 is found.
-    def check_for_job
-      resp = RestClient.get(
-        "#{config.orchestrator_address}/jobs/next"
-      )
-      job_data = JSON.parse(resp.body)
-
-      klass = case job_data['type']
-              when 'test_runner'
-                Jobs::TestRunnerJob
-              when 'representer'
-                Jobs::RepresenterJob
-              when 'analyzer'
-                Jobs::AnalyzerJob
-              else
-                raise "Unknown job: #{job_data['type']}"
-              end
-
-      klass.new(
-        job_data['id'],
-        job_data['language'],
-        job_data['exercise'],
-        job_data['source'],
-        job_data['container_version']
-      )
-    rescue RestClient::NotFound
-      nil
-    end
-
-    def handle_job(job)
-      ProcessJob.(job)
-      RestClient.patch(
-        "#{config.orchestrator_address}/jobs/#{job.id}",
-        {
-          status: job.status,
-          output: job.output
-        }
-      )
-      UploadMetadata.(job)
-    rescue StandardError => e
-      Log.("Error handling job", job:)
-      Log.(e.message, job:)
-      Log.(e.backtrace, job:)
     end
 
     def config
