+++ package.json
@@ -1,12 +1,15 @@
 {
   "devDependencies": {
-    "husky": "^4.2.5",
-    "pretty-quick": "^2.0.1",
-    "prettier": "^2.0.5"
+    "husky": "^8.0.3",
+    "prettier": "^2.0.5",
+    "pretty-quick": "^2.0.1"
   },
   "husky": {
     "hooks": {
       "pre-commit": "pretty-quick --staged && ./bin/rubocop-quick"
     }
+  },
+  "scripts": {
+    "prepare": "husky install"
   }
 }
